testcase/requirements.txt
pytest>=7.0.0
requests>=2.28.0
Pillow>=9.0.0
